+++ src/main/java/org/fairdatapoint/service/bootstrap/fixtures/SettingsFixture.java
@@ -20,23 +20,24 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.fairdatapoint.service.boostrap.fixtures;
+package org.fairdatapoint.service.bootstrap.fixtures;
 
 import lombok.Data;
-import org.fairdatapoint.entity.user.UserRole;
+import org.fairdatapoint.api.dto.search.SearchFilterDTO;
+import org.fairdatapoint.api.dto.settings.SettingsAutocompleteSourceDTO;
+import org.fairdatapoint.api.dto.settings.SettingsMetricDTO;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.UUID;
 
 @Data
-public class UserFixture {
-    private UUID uuid = UUID.randomUUID();
-    private String firstName = "";
-    private String lastName = "";
-    private String email;
-    private String password;
-    private UserRole role = UserRole.USER;
-    private List<String> apiKeyTokens = new ArrayList<>();
-    private List<SearchSavedQueryFixture> savedQueries = new ArrayList<>();
+public class SettingsFixture {
+    private String appTitle = "FAIR Data Point";
+    private String appSubtitle = "Metadata for Machines";
+    private Boolean pingEnabled = true;
+    private List<String> pingEndpoints = new ArrayList<>();
+    private Boolean autocompleteSearchNamespace = true;
+    private List<SettingsAutocompleteSourceDTO> autocompleteSources = new ArrayList<>();
+    private List<SettingsMetricDTO> metrics = new ArrayList<>();
+    private List<SearchFilterDTO> searchFilters = new ArrayList<>();
 }
