+++ src/main/java/org/fairdatapoint/service/search/query/SearchSavedQueryMapper.java
@@ -28,7 +28,7 @@ import org.fairdatapoint.api.dto.search.SearchSavedQueryChangeDTO;
 import org.fairdatapoint.api.dto.search.SearchSavedQueryDTO;
 import org.fairdatapoint.entity.search.SearchSavedQuery;
 import org.fairdatapoint.entity.user.UserAccount;
-import org.fairdatapoint.service.boostrap.fixtures.SearchSavedQueryFixture;
+import org.fairdatapoint.service.bootstrap.fixtures.SearchSavedQueryFixture;
 import org.fairdatapoint.service.user.UserMapper;
 import org.springframework.stereotype.Component;
