+++ src/main/java/org/fairdatapoint/service/settings/SettingsMapper.java
@@ -31,7 +31,7 @@ import org.fairdatapoint.config.properties.PingProperties;
 import org.fairdatapoint.config.properties.RepositoryConnectionProperties;
 import org.fairdatapoint.config.properties.RepositoryProperties;
 import org.fairdatapoint.entity.settings.*;
-import org.fairdatapoint.service.boostrap.fixtures.*;
+import org.fairdatapoint.service.bootstrap.fixtures.*;
 import org.springframework.stereotype.Component;
 
 import java.time.Instant;
